third_party/blink/web_tests/virtual/select-parsing-filtering-disabled/external/wpt/html/semantics/forms/the-select-element/customizable-select/select-input-parsing.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Input tags should parse inside select if there are no options first.
  assert_equals: expected "<button>button</button><input>" but got "<button>button</button>"
[FAIL] Input tags should parse inside select if nested in another tag
  assert_equals: expected "<div><input></div>" but got "<div></div>"
Harness: the test ran to completion.
